Added optionnal front openings
Added two front openings on each side of the camera opening. Each can be deactivated by setting frontOpening1 and/or frontOpening2 to 0. Those openings are similar to the camera one, and can receive a gasket and a part to close it. The default radius is the one used for the buttons (16mm+0.5 for easy fit), so those can be used for front buttons.
These openings can also be used with high intensity LED to provide some light to what is being observed.