updated electronics plate to move the pi back
The pi was too close to the camera module, i had not taken into account the µSD card.
Now it has been moved back. The current position makes the SD card annoying to remove, since the pi needs to be detached from the support.
However, this disposition leaves a maximum of room for the wifi module.